+++ lib/Views/RelatedMaps.jsx
@@ -65,12 +65,13 @@ function RelatedMaps(props) {
                          alt="NEII Viewer"/>
                 </a>
 
-                <a target="_blank" className={Styles.link} href="http://neiiviewer.nicta.com.au">NEII Viewer</a>
+                <a target="_blank" className={Styles.link} href="http://www.neii.gov.au/viewer/">NEII Viewer</a>
 
                 <p>
                     The National Environmental Information Infrastructure (NEII) is an information platform designed
                     to improve discovery, access and re-use of nationally significant environmental data. More
-                    information on the NEII is available here: [www.neii.gov.au](http://www.neii.gov.au)
+                    information on the NEII is available here: <a target="_blank" className={Styles.link} href="http://neii.gov.au/data-viewer">
+                    neii.gov.au/data-viewer</a>.
                 </p>
             </div>